Fix CallStack off-by-one, add call/byte-stack bounds checks, add register-indirect MOM/MOE addressing
CallStack.Call() wrote at stc_index then incremented; Return() read at
stc_index (already one past the last write) then decremented -- the first
RET after any CLL returned to address 0 instead of the pushed return
address. Fixed by decrementing before reading in Return(), matching Call's
write-then-increment convention. Zero prior test coverage of CLL/RET.

The call stack (int[255], static) had no depth check -- exceeding it threw
a raw IndexOutOfRangeException instead of a clean diagnostic. Added a bound
matching the house style (RAM.cs's "attempted to overwrite its own code"),
plus an underflow check on RET (empty call stack) and a Reset() the VM
constructor now calls, since the call stack is static/process-wide and was
silently leaking state across VM instances (confirmed empirically: it broke
an unrelated test purely from run order).

The byte stack (PSH/POP) lived in an isolated byte[256] buffer completely
disconnected from the VM's actual RAM, capped at 256 bytes with SP as a
byte -- unusable for real stack frames. Folded it into the shared address
space per Spec-Architecture.md: SP/SS widened to int, PSH/POP now go
through ram.SetByte/GetByte so the existing RAMLimit guard catches a stack
overflowing into code, plus a new explicit check for popping an empty
stack. SP/SS both default to the top of the VM's allocated RAM.

MOM/MOE only ever supported a literal address baked into the instruction --
no way to compute an address at runtime, making indexed/pointer-style
access impossible. Added register-indirect addressing by repurposing the
RegReg/ValReg AddressMode combinations these two opcodes never used (their
handlers ignored opMode for the address operand entirely): MOM AL, X now
stores at whatever address X currently holds, not a literal. No new
opcodes or encoding changes needed -- the compiler already emitted the
right AddressMode for a register operand in that position.

Also fixes DebugForm.cs's stack view, which indexed the now-removed
isolated buffer and hardcoded the old 256-byte range.

All fixes are TDD: each has a regression test that failed for the
documented reason before the fix and passes after. Full suite: 49/49.

+++ source/Artemis-VM/VM/VMextended.cs
@@ -21,8 +21,8 @@ namespace Artemis_IL
 
         /// <summary>
         /// Writes <paramref name="Content"/> into the register identified by <paramref name="Register"/>.
-        /// PC and IP are full-width, like X/Y, so they can address the entire RAM segment.
-        /// 8-bit registers (SS, AL, AH, BL, BH, CL, CH) are truncated to their low byte.
+        /// PC, IP and SS are full-width, like X/Y, so they can address the entire RAM segment.
+        /// 8-bit registers (AL, AH, BL, BH, CL, CH) are truncated to their low byte.
         /// Writing to the composite registers A (0xF4), B (0xF7), or C (0xFA) splits the 16-bit value
         /// across the corresponding L/H byte pair via <see cref="SetSplit"/>.
         /// Writing to SP (0xF2) is silently ignored — SP is managed automatically by stack operations
@@ -38,7 +38,7 @@ namespace Artemis_IL
                 IP = Content;
             //0xF2 (Stack Pointer) is read only
             else if (Register == (byte)0xF3)
-                SS = (byte)Content;
+                SS = Content;
             else if (Register == (byte)0xF4)
                 SetSplit('A', Content);
             else if (Register == (byte)0xF5)
@@ -79,9 +79,9 @@ namespace Artemis_IL
             else if (Register == (byte)0xF1)
                 return IP;
             else if (Register == (byte)0xF2)
-                return (int)SP;
+                return SP;
             else if (Register == (byte)0xF3)
-                return (int)SS;
+                return SS;
             else if (Register == (byte)0xF4)
                 return GetSplit('A');
             else if (Register == (byte)0xF5)
